+++ Oproto.FluentDynamoDb.UnitTests/Requests/FluentBuilderCompatibilityTests.cs
@@ -4,6 +4,7 @@ using AwesomeAssertions;
 using NSubstitute;
 using Oproto.FluentDynamoDb.Requests;
 using Oproto.FluentDynamoDb.Requests.Extensions;
+using Oproto.FluentDynamoDb.UnitTests.TestHelpers;
 
 namespace Oproto.FluentDynamoDb.UnitTests.Requests;
 
@@ -18,7 +19,7 @@ namespace Oproto.FluentDynamoDb.UnitTests.Requests;
 /// </summary>
 public class FluentBuilderCompatibilityTests
 {
-    private class TestEntity { }
+    private class TestEntity : TestEntityBase { }
     private readonly IAmazonDynamoDB _mockClient = Substitute.For<IAmazonDynamoDB>();
 
     #region Test all request builders work correctly with extension methods